MFC r315490, r315508:
Make the imx5 clocks driver work with vendor-supplied dts (which does not
supply the addresses for the DPLL register blocks) by hard-coding the
addresses in the driver source code.  Yes, this is just as bad an idea as
it sounds, but we have no choice.

Remove hardcoded bootverbose=1; imx53 systems boot using loader(8) and
users can set verbose if they want to.


git-svn-id: https://svn.freebsd.org/base/stable/11@323400 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f