MFC r316374, r316377:
Preserve the registers containing argc, argv, and return address values
passed in from u-boot across the call to self_reloc and any other early-init
code, and restore them before calling main().

Correct a comment... the stack used by ubldr is the same stack u-boot was
running on when it jumped to the ubldr entry point.  None of the arches
that use this code set up a different stack in their start.S routines.


git-svn-id: https://svn.freebsd.org/base/stable/11@323409 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f