MFC r320076:
Add a driver for the imx6 EPIT timer that can be used as the system
timecounter instead of the GPT timer, freeing up the more flexible GPT
hardware for other uses.  The EPIT driver is a standard (always in the
kernel) driver, and the existing GPT driver is now optional and included
only if you ask for device imx_gpt.


git-svn-id: https://svn.freebsd.org/base/stable/11@323418 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f