MFC r320928, r320929:
Support the "disable-over-current" property for imx ehci controllers.
It turns out that this is more than a power optization.  The OTG port
won't work on boards that have this property unless this setting is honored.
Also ensure that the usb phy device attaches before ehci.

Put an #ifdef notyet wrapper around a function that's not being used yet,
to avoid compile warnings.


git-svn-id: https://svn.freebsd.org/base/stable/11@323422 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f