MFC r321583, r321584:
Add a pair of convenience routines for doing simple "register" read/writes
on i2c devices, where the "register" can be any length.

Add support for tracking nested calls to iicbus_request/release_bus().
Usually it is sufficient to use iicbus_transfer_excl(), or one of the
higher-level convenience functions that use it, to reserve the bus for the
duration of each register access.  Occasionally it is important that a
series of accesses or read-modify-write operations must be done without any
other intervening access to the device, to prevent corrupting state.


git-svn-id: https://svn.freebsd.org/base/stable/11@323446 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f