MFC r323132-r323133
r323132:
The latest RPi firmware leaves secondary cores in a wait-for-event (WFE)
state to save power, so after writing the entry point address for a core to
the mailbox, use a dsb() to synchronize the execution pipeline to the data
written, then use an sev() to wake up the core.

Submitted by:	Sylvain Garrigues <sylgar@gmail.com>

r323133:
Change leading spaces to tabs, no functional change.


git-svn-id: https://svn.freebsd.org/base/stable/11@323471 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f