make-memstick.sh: use 'set -e' to abort if any step fails
Also remove the now-redundant error handling that was only for makefs.

This script was run on an older FreeBSD host that lacked efi-on-mbr
support in makefs. A warning was emitted on the console (from makefs)
but the script continued running and exited with 0.

MFC of r308171 (arm64) and r323178 (other architectures).


git-svn-id: https://svn.freebsd.org/base/stable/11@323476 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f