Unbreak netmap(4) support in ixgbe(4) after r315333: - Both ixgbe_netmap.c and ixv_netmap.c assumed a netmap(4) driver   newer than what's actually in stable/10. - Additionally, at the bottom line ixv_netmap.c did exactly the same   as ixgbe_netmap.c, i. e. used IXGBE_TDH() as appropriate for PFs   only instead of IXGBE_VFTDH() and tried to configure CRC stripping   although the corresponding registers aren't available to VFs in the   first place.
With these changes, the netmap(4) support in ixgbe(4) is in line
again with the code in sys/dev/netmap/ixgbe_netmap.h as of r295008.
Breakage reported by: Slawa Olhovchenkov

Just like r315333 that never existed in head, this is a direct commit
to stable/10. However, ixgbe(4) in head has a related bug in that it
assumes a netmap(4) driver API older than what's in head and also
does the wrong things for VFs as it uses the PF-only ixgbe_netmap.c
for both PFs and VFs in the first place.


git-svn-id: https://svn.freebsd.org/base/stable/10@323830 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f