MFC r323252:
Add a new getty/gettytab capability to generate an initial message dynamically.

This modification adds a new gettytab(5) option (iM) to specify a
program to run that will generate the initial (banner) message that is
displayed before the login prompt. Such a capability is useful when
dynamic information is needed in the banner message that cannot be
supplied by the set of % substitution sequences available in the "im"
option.

Sponsored by:	Dell EMC


git-svn-id: https://svn.freebsd.org/base/stable/11@323992 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f