Fix the return value from _Unwind_Backtrace() on arm.
If unwinding stops due to hitting the end of the call chain, the return
value is supposed to be _URC_END_OF_STACK; other values indicate internal
errors.  The return value from get_eit_entry() is now returned without
translating it to _URC_FAILURE, so that callers can see _URC_END_OF_STACK
when it happens.


git-svn-id: https://svn.freebsd.org/base/head@323998 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f