MFC r323183:
Make root_mount_rel(9) ignore NULL arguments, like it used to before r313351.
It would be better to fix API consumers to not pass NULL there - most of them,
such as gmirror, already contain the neccessary checks - but this is easier
and much less error-prone.

One known user-visible result is that it fixes panic on a failed "graid label".


git-svn-id: https://svn.freebsd.org/base/stable/11@324268 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f