MFC r323225:
Reflect realtime and idle priorities in ps(1) state flags, same like
we do for the usual nice values.  It could be argued that they should
use another set of indicators, since the underlying mechanism is
different, but they match the description in the manual page, and so
I think it's ok to not overcomplicate things.

PR:		81757


git-svn-id: https://svn.freebsd.org/base/stable/11@324270 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f