MFC r320803:
Fix "mount -uw /" when the filesystem type doesn't match.

This basically makes "mount -uw /" work when the filesystem
mounted on / is NFS, but the one configured in fstab(5) is UFS,
which can happen when you forget to modify fstab.

Note that the whole special case ("else if (argv[0][0] == '/'")
is probably not needed anyway.  I'll take a look at removing it
altogether; for now this is a minimally intrusive fix.


git-svn-id: https://svn.freebsd.org/base/stable/11@324273 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f