MFC r305508: r305508 (by markj):
Add some fail points to gmirror.

These are useful for testing changes to I/O error handling, and for
reproducing existing bugs in a controlled manner. The fail points are

    g_mirror_regular_request_read
    g_mirror_regular_request_write
    g_mirror_sync_request_read
    g_mirror_sync_request_write
    g_mirror_metadata_write

They all effectively allow one to inject an error value into the bio_error
field of a corresponding BIO request as it is being completed.


git-svn-id: https://svn.freebsd.org/base/stable/11@324402 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f